requirements.txt
appnope==0.1.0
backcall==0.1.0
backports.functools-lru-cache==1.5
bleach==3.0.0
certifi==2018.8.24
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.0.1
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
idna==2.7
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jaraco.functools==1.20
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
more-itertools==4.3.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.2
oauthlib==2.1.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
portend==2.3
prometheus-client==0.4.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
requests-oauthlib==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tempora==1.13
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
zc.lockfile==1.3.0
